Exercise Detail View
Clicking on an exercise now provides a full detailed view with stats, historical data, descriptions on how to do it, images, and a leaderboard (functionality TBD).
WIP is the ExerciseSearchView to incorporate this functionality while keeping the UI clean. Also multiselect when adding exercises to workouts and templates has been added