[BIG COMMIT] A whole lotta progress but a whole lotta nothing
Worked on the Leaderboard, following and followers. Everything broke, but this is stable. Profile page improvements, home feed isn't working anymore. Working on adding privacy constraints might've broken it. Need to work on displaying the workouts on the homepage. Then need to implement user discovery so the following function can be tested.