[BIG COMMIT] Community page implementation
Challenges can now be created, and users can join and leave them. Their workouts during the time window of the challenge count towards the challenge progress. So far, volume is the primary metric to index on. Teams is roughly implemented but still needs to be tested. The design of the Challenge details page can still be fixed up a bit. Need to test challenge completion and trophy earning, with a badge on the workout summary if a challenge is completed.